cleanup examples
