+++ examples/basic/prisma/schema.prisma
